Bypassed forces rule - lessen the footprint for logic on finding dependencies with force and version constraints for now
