+++ src/main/groovy/com/netflix/nebula/lint/rule/GradleLintRule.groovy
@@ -29,7 +29,6 @@ import org.codehaus.groovy.ast.expr.MapExpression
 import org.codehaus.groovy.ast.expr.MethodCallExpression
 import org.codehaus.groovy.ast.expr.PropertyExpression
 import org.codehaus.groovy.ast.expr.VariableExpression
-import org.codehaus.groovy.ast.stmt.BlockStatement
 import org.codehaus.groovy.ast.stmt.ExpressionStatement
 import org.codenarc.rule.AbstractAstVisitorRule
 import org.codenarc.rule.AstVisitor
@@ -81,10 +80,6 @@ abstract class GradleLintRule extends GroovyAstVisitor implements Rule {
     void visitAllprojectsGradleDependency(MethodCallExpression call, String conf, GradleDependency dep) {}
 
     void visitAnyGradleDependency(MethodCallExpression call, String conf, GradleDependency dep) {}
-
-    void visitAnyGradleDependencyWithAForceClosure(MethodCallExpression call, String conf, GradleDependency dep) {}
-
-    void visitAnyGradleDependencyWithVersionConstraint(MethodCallExpression call, String conf, GradleDependency dep, Map<String, List<String>> versionConstraints) {}
 
     void visitAnyObjectDependency(MethodCallExpression call, String conf, Object dep) {}
 
@@ -466,57 +461,8 @@ abstract class GradleLintRule extends GroovyAstVisitor implements Rule {
                             visitGradleDependency(call, methodName, dependency)
                         }
                         visitAnyGradleDependency(call, methodName, dependency)
-
-                        if (call.arguments.expressions
-                                .findAll { it instanceof ClosureExpression }
-                                .any { closureContainsForce(it as ClosureExpression) }) {
-                            visitAnyGradleDependencyWithAForceClosure(call, methodName, dependency)
-                        }
-
-                        List<Map<String, List<String>>> versionConstraints = call.arguments.expressions
-                                .findAll { it instanceof ClosureExpression }
-                                .collect { gatherVersionConstraints(it as ClosureExpression) }
-                        if (versionConstraints.size() > 0) {
-                            visitAnyGradleDependencyWithVersionConstraint(call, methodName, dependency, versionConstraints.first())
-                        }
-
                     }
                 }
-            }
-
-            private static Boolean closureContainsForce(ClosureExpression expr) {
-                return expr.code.statements.any {
-                    (it.expression instanceof BinaryExpression) &&
-                            ((BinaryExpression) it.expression).leftExpression?.variable == 'force' &&
-                            ((BinaryExpression) it.expression).rightExpression?.value == true
-                }
-            }
-
-            private static Map<String, List<String>> gatherVersionConstraints(ClosureExpression expr) {
-                def results = new HashMap<String, List<String>>()
-                expr.code.statements.findAll { st ->
-                    (st.expression instanceof MethodCallExpression) &&
-                            ((MethodCallExpression) st.expression)?.methodAsString == 'version' &&
-                            ((MethodCallExpression) st.expression)?.arguments?.findAll { arg ->
-                                arg instanceof ClosureExpression &&
-                                        arg?.code instanceof BlockStatement &&
-                                        arg?.code?.statements?.findAll { stmt ->
-                                            stmt?.expression instanceof MethodCallExpression &&
-                                                    stmt?.expression?.method instanceof ConstantExpression &&
-                                                    stmt?.expression?.arguments?.expressions?.findAll { expre ->
-                                                        expre instanceof ConstantExpression &&
-                                                                !expre?.value?.equals(null)
-                                                    }
-                                                            ?.each {
-                                                                String method = stmt?.expression?.methodAsString
-                                                                List values = stmt?.expression?.arguments?.expressions?.collect { it.value as String }
-
-                                                                results.put(method, values)
-                                                            }
-                                        }
-                            }
-                }
-                return results
             }
 
             private void visitMethodCallInPlugins(MethodCallExpression call) {
